+++ docs/changelog.md
@@ -45,6 +45,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: use provided priority or else fall back to priority of parent issue
 * i18n: change label for multiple sub-tasks dialog
 * feat: display input string and potential input errors after submit
+* feat: support explicit input for issue type of the created sub-tasks
 
 [Unreleased]: https://github.com/codescape/jira-multiple-subtasks/compare/1.0.0...HEAD
 [1.0.0]: https://github.com/codescape/jira-multiple-subtasks/tree/1.0.0